Create CS 140 Calendar part 2
This was a part of the large quarter-long project in CS 140 (intro to java) in which we created a calendar that would display ASCII art and a list of dates.  This was the second part of the calendar, by the end we were able to jump to certain dates and it was possible to change the size of the calendar.